# libs/ocr_common/ocr_common/web/schemas.py
"""Pydantic models and helpers of the response envelope and of the documented examples."""

from typing import Any, Literal

from pydantic import BaseModel, Field

from ocr_common.web.envelope import envelope

REQUEST_ID_EXAMPLE = "REQ_9cb01af2-493d-446d-b191-af120333f6d0"

Stage = Literal["OCR", "STRUCTURING", "SCORING"]
JobState = Literal["PROCESSING", "DONE", "FAILED"]


class SuccessEnvelope(BaseModel):
    """Base of every success response model; subclasses declare `data`."""

    status_code: int = Field(200, description="Same as the HTTP status code", examples=[200])
    status_desc: str = Field("OK", description="Reason phrase of `status_code`", examples=["OK"])
    message: str = Field("Success", description="Human-readable outcome", examples=["Success"])
    errors: None = Field(None, description="Always null on success")
    request_id: str | None = Field(
        None,
        description=(
            "The request_id this response belongs to: the one sent in the form / body / path, otherwise the "
            "`X-Request-ID` header, otherwise one generated by the service"
        ),
        examples=[REQUEST_ID_EXAMPLE],
    )


class ErrorResponse(BaseModel):
    """Every error response."""

    status_code: int = Field(..., description="Same as the HTTP status code", examples=[400])
    status_desc: str = Field(..., description="Reason phrase of `status_code`", examples=["Bad Request"])
    message: str = Field(..., description="Human-readable cause; safe to log", examples=["Uploaded file is empty"])
    data: None = Field(None, description="Always null on error")
    errors: str = Field(
        ...,
        description=(
            "Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message names a different "
            "field every time); for every other status it repeats `message`"
        ),
        examples=["Uploaded file is empty"],
    )
    request_id: str | None = Field(
        None, description="Present when the error is tied to a known request_id", examples=[REQUEST_ID_EXAMPLE]
    )


class JobAccepted(BaseModel):
    """`data` of `POST /v1/<stage>/jobs`."""

    request_id: str = Field(..., description="Echo of the submitted request_id", examples=[REQUEST_ID_EXAMPLE])
    stage: Stage = Field(..., description="Pipeline stage that accepted the job", examples=["OCR"])
    status: JobState = Field(
        ...,
        description=(
            "`PROCESSING` for a newly accepted job. When `duplicate` is true: the current status of the job "
            "that already exists for this request_id"
        ),
        examples=["PROCESSING"],
    )
    duplicate: bool = Field(
        False,
        description=(
            "true: this request_id was already accepted before and the work is NOT run again (idempotent). "
            "Exception: a job whose status is `FAILED` is re-claimed and run again, and answers `duplicate: false`"
        ),
        examples=[False],
    )


class JobAcceptedResponse(SuccessEnvelope):
    """Envelope of `POST /v1/<stage>/jobs`."""

    status_code: int = Field(202, description="Same as the HTTP status code", examples=[202])
    status_desc: str = Field("Accepted", description="Reason phrase of `status_code`", examples=["Accepted"])
    message: str = Field("Accepted", description="Human-readable outcome", examples=["Accepted"])
    data: JobAccepted


class JobStatusBase(BaseModel):
    """Common fields of `GET /v1/<stage>/jobs/{request_id}`; each stage adds its `result`."""

    request_id: str = Field(..., description="request_id of the pipeline run", examples=[REQUEST_ID_EXAMPLE])
    stage: Stage = Field(..., description="Pipeline stage this job belongs to", examples=["OCR"])
    status: JobState = Field(
        ...,
        description="`PROCESSING` -> `DONE` | `FAILED`. A `FAILED` job may go back to `PROCESSING` when resubmitted",
        examples=["DONE"],
    )
    error_message: str | None = Field(
        None,
        description="Why the stage failed; null unless `status` is `FAILED`. Same text as the FAILED callback",
        examples=[None],
    )
    created_at: str = Field(
        ..., description="When the job was first accepted (ISO 8601, UTC)", examples=["2026-09-18T04:00:00+00:00"]
    )
    updated_at: str = Field(
        ..., description="Last status change (ISO 8601, UTC)", examples=["2026-09-18T04:00:01+00:00"]
    )
    pipeline_name_sequence: list[str] | None = Field(
        None,
        description=(
            "The services this request runs, as the job was submitted with them; the last one ends the request. "
            "Null for a job submitted without it: the full pipeline"
        ),
        examples=[["guardrails", "ekstraksi", "structuring", "scoring"]],
    )


PIPELINE_SEQUENCE_DESCRIPTION = (
    "The services this request runs, in order, from the central orchestrator (forwarded by the orchestrator "
    "and each stage): `guardrails`, `ekstraksi`, `structuring`, `scoring`, guardrails optional at the front "
    "and the end cut off, never one skipped in the middle. The last one ends the request: its result is the "
    "answer, as it is, and nothing is handed on. Omitted: the full pipeline"
)


class HealthResponse(BaseModel):
    """`GET /health`."""

    status: str = Field(..., description="Always `healthy` while the process is alive", examples=["healthy"])
    version: str = Field(..., description="Service version", examples=["1.0.0"])
    device: str = Field(..., description="Where this service computes", examples=["cpu"])
    backends: dict[str, str] = Field(
        ...,
        description="Active implementation per concern (model backend, storage). Useful to confirm a deployment",
        examples=[{"ekstraksi": "paddle", "storage": "postgres"}],
    )


class ReadyResponse(BaseModel):
    """`GET /ready`."""

    status: Literal["ready", "not_ready"] = Field(
        ..., description="`ready` (HTTP 200) when every check is `ok`, else `not_ready` (HTTP 503)", examples=["ready"]
    )
    checks: dict[str, Literal["ok", "failed"]] = Field(
        ...,
        description="One entry per REQUIRED dependency of this service; empty when it has none",
        examples=[{"database": "ok"}],
    )


def error(
    status_code: int,
    description: str,
    message: str,
    *,
    request_id: str | None = None,
    errors: str | None = None,
) -> dict:
    """A documented error response with its own envelope example."""
    return {
        "model": ErrorResponse,
        "description": description,
        "content": {
            "application/json": {
                "example": envelope(status_code, message, None, request_id, errors=errors or message),
            }
        },
    }


def success_examples(description: str, **named: tuple[str, Any]) -> dict:
    """A documented 200 response with named envelope examples."""
    return {
        "description": description,
        "content": {
            "application/json": {
                "examples": {name: {"summary": summary, "value": value} for name, (summary, value) in named.items()}
            }
        },
    }


UNAUTHORIZED = error(401, "Missing or invalid X-API-Key", "Invalid or missing API key")