# services/scoring/openapi.yaml
openapi: 3.1.0
info:
  title: OCR Kartu Keluarga Scoring API
  description: 'Pipeline step 4 (last) for Indonesian Kartu Keluarga: P(correct) per contract field, after
    fusing the two structuring scores. There is no document score and no approve/reject decision -- the
    threshold belongs to the caller. **Async pipeline:** the structuring service POSTs /v1/scoring/jobs
    and gets 202; this service scores in the background, stores the result, and POSTs the stage callback
    carrying the final result to the orchestrator. /v1/scoring/confidence is the same work, synchronous.
    All endpoints except /health require an X-API-Key header.


    ## Conventions (the same for every nlm-k2 service)


    **Authentication.** Every endpoint except `/health`, `/ready` and `/metrics` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.


    **Errors.** `errors` is machine-readable: `VALIDATION_ERROR` for `422`, otherwise the same text as

    `message`. `400` = the request or the document is unusable (do not retry unchanged). `401` = API key.

    `404` = unknown request_id. `409` = request_id already processed (legacy contract only). `422` = a

...

...


...

...

...


...

...

...
servers:
...
paths:
...
  /v1/scoring/jobs/{request_id}:
    get:
      tags:
      - Pipeline
      summary: Status and result of the scoring stage
      description: 'Status of this stage only and, once `DONE`, the two confidences plus the exact payload
        that was scored. Internal: the orchestrator reads it (while it waits, and for its `GET /v1/extract-ocr/{request_id}`);
        the central orchestrator receives the final result in the callback. Also useful to audit a score.'
      operationId: getScoringJob
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: The job exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoringJobStatusResponse'
              examples:
                done:
                  summary: Finished
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: DONE
                      result:
                        nomor_kk: 0.9806
                        name_confidence: 0.9948
                        payload:
                          structuring: '{...the §7.3 structuring result, whole...}'
                          guardrail_probability: 0.0287
                          guardrail_verdict: accepted
                          avg_doc_score: 0.814
                          min_doc_score: 0.2822
                          text_regions_count: 177
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                failed:
                  summary: Failed
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: FAILED
                      error_message: 'Unsupported document_type: ktp. Supported: [''kk'']'
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No scoring job for this request_id
          content:
            application/json:
              example:
                status_code: 404
                status_desc: Not Found
                message: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                errors: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'path.request_id: Field required'
                errors: VALIDATION_ERROR
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
      description: Every error response.
...
    ScoringJobResult:
      properties:
        document_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
          description: Document type the fields were read as
          examples:
          - kk
        fields:
          additionalProperties:
            anyOf:
            - type: number
            - type: 'null'
          type: object
          title: Fields
          description: '`nomor_kk` and `nama_kepala_keluarga`. Null for a field whose value is empty'
          examples:
          - nama_kepala_keluarga: 0.8871
            nomor_kk: 0.9412
        anggota_keluarga:
          items:
            additionalProperties:
              anyOf:
              - type: number
              - type: 'null'
            type: object
          type: array
          title: Anggota Keluarga
          description: Seven scores per member, positionally aligned with the structuring result. A length
            mismatch is a defect in this pipeline, not a property of the document, and is rejected rather
            than truncated
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Identity of the trust model
          examples:
          - kk-trust-isotonic-v1
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
          description: 'Exactly what was scored, as an audit trail: §8.3 requires the numbers to be reproducible
            without re-running the pipeline, which is only true if the input is stored with the output'
        thresholds:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Thresholds
          description: The confidence above which a field of that name was correct on every held-out sample
            -- one per scored field, keyed by INTERNAL name. They travel in the result rather than in
            configuration because they are a property of the trained model, not of the deployment, and
            because this is what makes the orchestrator's `auto` and the outcome row's `auto` identical
            by construction instead of by keeping two env vars in step. A field absent here falls back
            to `FIELD_CONFIDENCE_THRESHOLD`
          examples:
          - ayah: 0.9353
            nik: 0.9637
            pendidikan: 0.9929
        bin_edges:
          anyOf:
          - additionalProperties:
              items:
                type: number
              type: array
            type: object
          - type: 'null'
          title: Bin Edges
          description: 'Bin boundaries low to high, under `fields` and `anggota_keluarga` -- the two model
            families have their own. Deliberately NOT equal width: the top bin is cut exactly where held-out
            precision reached 100%, which no fixed grid lands on. Null falls back to ten equal-width bins'
          examples:
          - anggota_keluarga:
            - 0.0
            - 0.631
            - 0.776
            - 0.839
            - 0.899
            - 0.925
            - 0.943
            - 0.965
            - 0.979
            - 0.9874
            - 1.0
      additionalProperties: true
      type: object
      title: ScoringJobResult
      description: 'Trust model output: P(each extracted field is correct), fused and calibrated.'
    ScoringJobStatus:
      properties:
        request_id:
          type: string
          title: Request Id
          description: request_id of the pipeline run
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        stage:
          type: string
          const: SCORING
          title: Stage
          description: Always `SCORING` on this service
          default: SCORING
          examples:
          - SCORING
        status:
          type: string
          enum:
          - PROCESSING
          - DONE
          - FAILED
          title: Status
          description: '`PROCESSING` -> `DONE` | `FAILED`. A `FAILED` job may go back to `PROCESSING`
            when resubmitted'
          examples:
          - DONE
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
          description: Why the stage failed; null unless `status` is `FAILED`. Same text as the FAILED
            callback
          examples:
          - null
        created_at:
          type: string
          title: Created At
          description: When the job was first accepted (ISO 8601, UTC)
          examples:
          - '2026-09-18T04:00:00+00:00'
        updated_at:
          type: string
          title: Updated At
          description: Last status change (ISO 8601, UTC)
          examples:
          - '2026-09-18T04:00:01+00:00'
        pipeline_name_sequence:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Pipeline Name Sequence
          description: 'The services this request runs, as the job was submitted with them; the last one
            ends the request. Null for a job submitted without it: the full pipeline'
          examples:
          - - guardrails
            - ekstraksi
            - structuring
            - scoring
        result:
          anyOf:
          - $ref: '#/components/schemas/ScoringJobResult'
          - type: 'null'
          description: Confidences once `status` is `DONE`; null otherwise
      type: object
      required:
      - request_id
      - status
      - created_at
      - updated_at
      title: ScoringJobStatus
    ScoringJobStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/ScoringJobStatus'
      type: object
      required:
      - data
      title: ScoringJobStatusResponse
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...